+++ packages/@expo/cli/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### 💡 Others
 
+- Move server action env to `@expo/metro-runtime`.
 - Update `@urql/core` and related dependencies and remove `graphql` dependency. ([#32065](https://github.com/expo/expo/pull/32065) by [@kitten](https://github.com/kitten))
 
 ## 0.19.13 — 2024-10-31