+++ packages/@expo/metro-runtime/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- Move server action env to `@expo/metro-runtime`.
+
 ## 4.0.0-preview.0 — 2024-10-22
 
 ### 🎉 New features